requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
gmssl==3.2.2
idna==3.10
macholib==1.16.3
packaging==24.2
pycryptodomex==3.21.0
PyExecJS==1.5.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
requests==2.32.3
setuptools==75.5.0
six==1.16.0
urllib3==2.2.3
